Fix for the map 3d opacity

+++ bundles/net.tourbook.common/src/net/tourbook/common/color/Map3GradientColorManager.java
@@ -98,9 +98,9 @@ public class Map3GradientColorManager {
    private static final RGB    DEFAULT_RGB                     = new RGB(0xFF, 0x8B, 0x8B);
 
    // opacity
-   public static final int   OPACITY_MIN     = 0;
-   public static final int   OPACITY_MAX     = 100;
-   public static final float OPACITY_DEFAULT = 100;
+   public static final int OPACITY_MIN     = 0;
+   public static final int OPACITY_MAX     = 100;
+   public static final int OPACITY_DEFAULT = 100;
 
    /**
     * Define 3D map default colors.
